# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Nexus Messaging API
  description: |
    Nexus is a high-concurrency WebSocket messaging service with a CQRS-inspired architecture.

    **Architecture Overview:**
    - **Ingress Tier** (this API): Stateless HTTP endpoints for sending messages, managing groups, user profiles, and authentication.
    - **Egress Tier**: Stateful WebSocket connections for real-time event delivery (MSG_NEW, MSG_UPDATED, MSG_STATUS, USR_JOINED, USR_LEFT, TYP_INDICATOR).
    - **NATS Cluster**: Internal message bus connecting Ingress and Egress tiers.

    **Authentication:**
    All protected endpoints require a JWT token. Nexus does not provide registration or login flows —
    developers implement their own auth service and use the internal API to sync users into Nexus.
    Pass the token in the `Authorization: Bearer <token>` header.
  version: "1.0.0"
  contact:
    name: Nexus Team
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with claims:
        - `sub` (string): user ID
        - `tid` (string): tenant/app ID
        - `iat` (int): issued at
        - `exp` (int): expiration (24h from issue)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Machine-readable error code
      required:
        - success
        - message
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint64
          description: Snowflake message ID
          example: 420158066278400
        app_id:
          type: string
          example: "myapp"
        recipient_type:
          type: string
          enum: [group, user]
          description: "Type of recipient: 'group' for group messages, 'user' for private messages"
          example: "group"
        recipient_id:
          type: string
          description: Target identifier (group ID when recipient_type is group, user ID when user)
          example: "nexus_dev_team"
        shard_id:
          type: integer
          example: 42
        content:
          type: object
          description: Structured message content (JSON)
          example: { "type": "text", "text": "Hello, world!" }
        sender_type:
          type: string
          enum: [user]
          description: "Type of sender (currently only 'user')"
          example: "user"
        sender_id:
          type: string
          description: Sender identifier (user ID when sender_type is user)
          example: "420158066212864"
        created_at:
          type: string
          format: date-time
          example: "2026-03-24T10:30:00Z"
...
paths:
...
  /internal/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register or retrieve a user (internal)
      description: |
        Service-to-service endpoint for syncing users from an external auth service into Nexus.
        Creates the user if they don't exist, or retrieves them if they do, and returns a JWT.
        Authenticated via `X-Internal-Key` header.
      operationId: registerUser
      parameters:
        - in: header
          name: X-Internal-Key
          required: true
          schema:
            type: string
          description: Internal API key configured via `http.internal_api_key`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                name:
                  type: string
                  description: Display name (optional, set on newly created users)
                  example: "John Doe"
                app_id:
                  type: string
                  description: Tenant ID (defaults to server's configured app ID if omitted)
                  example: "myapp"
      responses:
        "200":
          description: User registered or retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token (valid for 24 hours)
                        example: "eyJhbGciOiJIUzI1NiIs..."
                      user_id:
                        type: string
                        example: "420158066212864"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_json:
                  value: { "success": false, "message": "invalid_json" }
                email_required:
                  value: { "success": false, "message": "email_required" }
        "401":
          description: Invalid or missing internal API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                { "success": false, "message": "invalid_internal_key" }
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                { "success": false, "message": "register_failed" }
...
  /v1/history:
    get:
      tags: [History]
      summary: Get message history
      description: |
        Fetches message history for a group or private conversation. Returns messages
        with ID greater than `since_id`, ordered by ID ascending, up to `limit` messages.
        Use `group_id` for group history or `user_id` for private conversation history.
        Exactly one must be provided. For group history, the user must be a member.
      operationId: getHistory
      security:
        - BearerAuth: []
      parameters:
        - name: group_id
          in: query
          required: false
          description: Group ID to fetch history for (mutually exclusive with user_id)
          schema:
            type: string
          example: "nexus_dev_team"
        - name: user_id
          in: query
          required: false
          description: User ID to fetch private conversation history with (mutually exclusive with group_id)
          schema:
            type: string
          example: "420158066212864"
        - name: since_id
          in: query
          required: false
          description: Return only messages with ID greater than this value (cursor-based pagination)
          schema:
            type: integer
            format: uint64
            default: 0
          example: 420158066278400
        - name: limit
          in: query
          required: false
          description: Maximum number of messages to return (1-100, default 25)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          example: 25
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      messages:
                        type: array
                        items:
                          $ref: "#/components/schemas/Message"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                group_id_or_user_id_required:
                  value: { "success": false, "message": "group_id_or_user_id_required" }
                invalid_limit:
                  value: { "success": false, "message": "invalid_limit" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a member of the group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                { "success": false, "message": "not_a_group_member" }
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                { "success": false, "message": "history_failed" }